+++ release-notes.md
@@ -1,2 +1,7 @@
+`1.0.1` - 2023-06-15
+* Add HUD info on token mouseover
+* Add Wifi dead zone support
+* Add NPC XP points and Treason Stars
+
 `1.0.0` - 2023-06-09
 * Initial release! yay!